+++ src/utils/split.ts
